+++ processor/src/test/resources/bad_input/com/example/arez/ComputedOnPrivateComponent.java
@@ -1,0 +1,25 @@
+package com.example.arez;
+
+import arez.annotations.Computed;
+import javax.annotation.Nullable;
+import react4j.ReactNode;
+import react4j.annotations.ReactComponent;
+import react4j.arez.ReactArezComponent;
+
+@ReactComponent
+abstract class ComputedOnPrivateComponent
+  extends ReactArezComponent
+{
+  @Nullable
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+
+  @Computed
+  private boolean isActive()
+  {
+    return true;
+  }
+}