+++ processor/src/test/resources/bad_input/com/example/arez/MemoizeOnPrivateComponent.java
@@ -1,0 +1,25 @@
+package com.example.arez;
+
+import arez.annotations.Memoize;
+import javax.annotation.Nullable;
+import react4j.ReactNode;
+import react4j.annotations.ReactComponent;
+import react4j.arez.ReactArezComponent;
+
+@ReactComponent
+abstract class MemoizeOnPrivateComponent
+  extends ReactArezComponent
+{
+  @Nullable
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+
+  @Memoize
+  private String getIcon( String key )
+  {
+    return null;
+  }
+}